Update to latest framework versions

+++ client/src/containers/pieces/ViewPiece.tsx
@@ -2,11 +2,11 @@ import * as React from 'react';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
 import ViewablePiece from '../../components/pieces/ViewablePiece';
-import { useHistory } from 'react-router-dom';
-import { useAuth0 } from '../../contexts/auth0-context';
+import { useNavigate } from 'react-router-dom';
+import { useAuth0 } from '@auth0/auth0-react';
 
 function ViewPiece() {
-  let history = useHistory();
+  let navigate = useNavigate();
   let { tourId, pieceId } = useParams();
   const { isAuthenticated, getIdTokenClaims } = useAuth0();
 
@@ -24,7 +24,7 @@ function ViewPiece() {
   }, [tourId, pieceId]);
 
   const onEdit = async (e: React.MouseEvent): Promise<void> => {
-    history.push(`/tours/${tourId}/pieces/${pieceId}/edit`);
+    navigate(`/tours/${tourId}/pieces/${pieceId}/edit`);
   };
 
   const onDelete = async (e: React.MouseEvent): Promise<void> => {
@@ -34,13 +34,16 @@ function ViewPiece() {
     setDeleteSuccess(deleteSuccess);
     setLoading(false);
     setTimeout(() => {
-      history.push(`/tours/${tourId}/pieces`);
+      navigate(`/tours/${tourId}/pieces`);
     }, 500);
   };
 
   const deletePiece = async () => {
     try {
       const accessToken = await getIdTokenClaims();
+      if (!accessToken) {
+        throw new Error("Access token error");
+      }
       const response = await fetch(`${process.env.REACT_APP_SERVER_BASE_URL}/api/tours/${tourId}/pieces/${pieceId}`, {
         method: 'delete',
         headers: new Headers({
